Move specific tests into separate files.
